requirements_test.txt
# Test framework
pytest-django==2.9.1
pytest-cov==2.2.1

# Freezes time in tests
freezegun==0.3.7
